C16 - CORS configuration and more
Added the necessary CORS configuration so the code can be deployed to the online servers.
Added validation to the request bodies of the services.
Added an endpoint to get the information of a service reservation with the internal details of the reservation.
Added the option in the fillDB.py script to use existing users so it can be used in the online servers. Also modified the values generated for the alergies, discapacities and medicines for a person.
Added an endpoint to check if a user has a reservation with a pending service reservation of transportation.
Fun fact: This is likely the last mandatory commit I'll need for the MVP, the following commits will be life quality improvements to the project, but the core of the project is finally done.

+++ src/main/java/com/caritas/backend/dev/DevController.java
@@ -5,6 +5,7 @@ import java.util.UUID;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.caritas.backend.common.errors.BadRequestException;
@@ -102,7 +103,7 @@ public class DevController {
     }
 
     @PostMapping("/wipe")
-    public void wipeData() {
+    public void wipeData(@RequestParam(defaultValue = "false") String prod) {
         serviceReservationRepository.deleteAll();
         personReservationRepository.deleteAll();
         reservationRepository.deleteAll();
@@ -110,7 +111,9 @@ public class DevController {
         hostelRepository.deleteAll();
         serviceRepository.deleteAll();
         personRepository.deleteAll();
-        userRepository.deleteAll();
+        if (prod == "false") {
+            userRepository.deleteAll();
+        }
 
         bathReservationRepository.deleteAll();
         breakfastReservationRepository.deleteAll();
